Using SSR for the main page instead of client side for better PLT
