+++ components/ProjectCarousel.js
@@ -1,4 +1,6 @@
-import React, { useState, useEffect } from 'react';
+"use client";
+
+import React, { useState } from 'react';
 import ProjectCard from './ProjectDisplayCard';
 
 // Import Swiper React components
